requirements.txt
discord~=2.3.2
python-dotenv~=1.0.1
requests~=2.32.3
pydantic~=2.7.3
SQLAlchemy~=2.0.30
